Fix #31
This commit replaces the use of the `std::make_tuple` function with an
explicit call to the c++ tuple constructor. The old version did relay
the implementation defined argument evaluation order of function calls.
This broke on g++. See for example this question for more details: https://stackoverflow.com/questions/14056000/how-to-avoid-undefined-execution-order-for-the-constructors-when-using-stdmake