Provide support for solidity in `serde_generate` (#61)
## Summary

Solidity is the Javascript of blockchains, therefore having a way to
support solidity in one of the codes is quite important.

Solidity has a number of limitations that force some adaptations:
* The only enum supported are of the elementary types, therefore, the
more complex enums are supported as a "choice" entry in the code.
* There is no Tuple, so again a struct is introduced.
* There is no Option type, therefore, we again introduce a struct
construction.
* The TupleArray [T; N] are also encoded as struct with the length being
part of the typename.

There are also some more mundane restrictions where we choose to simply
panic with an explicit error if this occurs in the input code:
* No floating point.
* No empty struct
* No empty enum
* Enum with at most 256 entries.
* No recursive structure like trees.

The main implication of the use of struct to represent nested structure
is that it breaks the design of `ContainerFormat`/ `Format`. So, we need
to introduce the recursive type `SolFormat` type that provides support
for this.

Other design choices:
* No runtime is used. There is a way to have libraries in solidity, but
it complicates the development since the library needs to be deployed
first followed by the other contract.
* Everything uses the "memory" location. I tried to use the more
efficient "calldata" without success. Not sure if that is possible.
* The `get_keywords` is a standard function instead of being a static
one as in the OCaml generator. I think it does not matter.
* Solidity has the `abi.encode` / `abi.decode` that works with padded
serialization. It has an `abi.encodePacked` that provides a compact
serialization, but it has no `abi.decodePacked`. Added to that the BCS
serialization of 42_u32 is [42,0,0,0] but the solidity serialization is
[0,0,0,42]. Therefore hand-crafted code has been written for processing
all primitive types.
* No assembly is used for the operation. There is room for improvement
here.

## Test Plan

The code is tested by running an EVM virtual machine. This is nicely
provided by the REVM crate.
The only external files used are the temporary ones created during the
test which makes the development smooth.