fix goland CI (#68)
## Summary

Fix CI for Golang after a change in the default behavior.

I tried calling `go mod tidy` but it complains about name collision in
some cases so just reverting to the previous behavior.

However `-mod=mod` is not working well for golang 1.15 so I upgraded CI
to 1.22.

See also
https://stackoverflow.com/questions/66469396/go-module-is-found-and-replaced-but-not-required#comment117706021_66470269

## Test Plan

CI