Make Dart linter happy by adapting to its rules (#71)
## Summary

This PR makes Dart linter happy by adapting to its rules.

While applying `serde-reflection` to
[`rinf`](https://github.com/cunarist/rinf), warnings were generated by
the Dart language server. This PR addresses all of them.

Below are some of the warnings that were shown in VScode.


![image](https://github.com/user-attachments/assets/ec926852-2a1d-4277-9e5c-c65bd7139e4d)

![image](https://github.com/user-attachments/assets/3510dc58-c828-4687-91ba-e267602e30f8)


## Test Plan

The runtime behavior shouldn't change. If the CI works, then everything
would be fine.